+++ CHANGES.md
@@ -3,6 +3,8 @@ Change Log
 
 ### 1.66.0 - 2020-02-03
 
+##### Additions :tada:
+
 * Added `Globe.showSkirts` to support the ability to hide terrain skirts when viewing terrain from below the surface. [#8489](https://github.com/AnalyticalGraphicsInc/cesium/pull/8489)
 
 ### 1.65.0 - 2020-01-06